# openapi.json
{
  "components": {
    "schemas": {
      "Health": {
        "properties": {
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "Health",
        "type": "object"
      },
      "Me": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "name",
          "avatar_url",
          "email",
          "degraded"
        ],
        "title": "Me",
        "type": "object"
      },
      "NotReady": {
        "description": "Deliberately not a problem document: an orchestrator reads this and wants\nthe reason in a fixed place.",
        "properties": {
          "postgres": {
            "title": "Postgres",
            "type": "string"
          },
          "status": {
            "const": "not_ready",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "postgres"
        ],
        "title": "NotReady",
        "type": "object"
      },
      "Problem": {
        "additionalProperties": true,
        "description": "Extra members are allowed: `last_admin` carries `scopes`, for instance.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
      "Ready": {
        "properties": {
          "status": {
            "const": "ready",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "Ready",
        "type": "object"
      },
      "RegisterUrl": {
        "properties": {
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "url"
        ],
        "title": "RegisterUrl",
        "type": "object"
      },
      "ServerTime": {
        "properties": {
          "now": {
            "format": "date-time",
            "title": "Now",
            "type": "string"
          }
        },
        "required": [
          "now"
        ],
        "title": "ServerTime",
        "type": "object"
      },
      "SessionInfo": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current": {
            "title": "Current",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          }
        },
        "required": [
          "id",
          "created_at",
          "last_seen_at",
          "ip",
          "user_agent",
          "current"
        ],
        "title": "SessionInfo",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Contests on top of Forgejo, Woodpecker and Garage",
    "title": "Unicon API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/callback": {
      "get": {
        "operationId": "completeLogin",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "default": "",
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "default": "",
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "Land back from the forge",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "get": {
        "operationId": "startLogin",
        "parameters": [
          {
            "description": "A path on this site to land on",
            "in": "query",
            "name": "next",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A path on this site to land on",
              "title": "Next"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "Start a login through the forge",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "End this session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register-url": {
      "get": {
        "operationId": "getRegisterUrl",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterUrl"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "Where to create a forge account, if it is open",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/me": {
      "delete": {
        "operationId": "deleteMe",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "Delete the forge account",
        "tags": [
          "me"
        ]
      },
      "get": {
        "operationId": "getMe",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "The signed-in person",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/deactivate": {
      "post": {
        "operationId": "deactivateMe",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "Deactivate the forge account",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/sessions": {
      "delete": {
        "operationId": "revokeAllMySessions",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "Sign out everywhere, including here",
        "tags": [
          "me"
        ]
      },
      "get": {
        "operationId": "listMySessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionInfo"
                  },
                  "title": "Response Listmysessions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "This person's sessions",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/sessions/{session_id}": {
      "delete": {
        "operationId": "revokeMySession",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "End one session",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/time": {
      "get": {
        "operationId": "getServerTime",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerTime"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "Server time in UTC",
        "tags": [
          "time"
        ]
      }
    },
    "/healthz": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "Is the process up",
        "tags": [
          "health"
        ]
      }
    },
    "/readyz": {
      "get": {
        "operationId": "getReadiness",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ready"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotReady"
                }
              }
            },
            "description": "Postgres did not answer"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as an RFC 9457 problem document. `code` names it."
          }
        },
        "summary": "Is Postgres reachable",
        "tags": [
          "health"
        ]
      }
    }
  }
}